+++ packages/app/src/components/Loading.vue
@@ -1,0 +1,20 @@
+
+<template>
+  <div id="loading">Loading</div>
+</template>
+
+<script>
+export default {
+  name: "Loading"
+};
+</script>
+
+<style scoped>
+#loading {
+  position: absolute;
+  height: 200px;
+  width: 100%;
+  margin: auto;
+  top: 200px;
+}
+</style>
